+++ src/sentry/static/sentry/app/components/events/interfaces/threads/trimFilename.tsx
